grids
git-svn-id: http://alageospatialportal.googlecode.com/svn/trunk@457 73def43e-0e3b-11df-a5af-b38b39d06e06

+++ alaspatial/src/main/java/org/ala/spatial/util/Grid.java
@@ -39,7 +39,7 @@ public class Grid { //  implements Serializable
     public double xres, yres;
     public String datatype;
     // properties
-    public float minval, maxval;
+    public double minval, maxval;
     byte nbytes;
     String filename;
     float [] grid_data = null;
